Fixed issues related to skipping mzXML file loads.
We do not anticipate the user to fill out the mzXML file name column in the Peak Annotation Details sheet (though it's there in case the name doesn't match the sample header or when disambiguation is necessary in the face of same-named files.  This means that the skipping of files for skipped sample headers needs to account for the fact that the files can have the same name and are thus stored in a dict that includes the directories.  The way it was written, it would only skip files whose paths were provided in the Peak Annotation Details sheet, thus this PR adds code that makes an assumption that if the sample header matches the mzXML file name, and the number of files matches the number of skipped rows in the Peak Annotation Details sheet, then the load of those files should be skipped.

This code also accounts for the possibility that paths might be provided for not all matching sample headers, or when not all sample headers are skipped.

Details:

- Added exception class: MzXMLSkipRowError
- Changed the value at the end of the skip_msrunsample_by_mzxml dict to indicate the number of matching files (instead of bool) so that we could compare the non-matching actual files with the number of rows that are skipped for that sample header.
- When the number of files after the skip filter is zero, we no longer try to look up a matching sample.  This fixes a bug where we were getting DoesNotExist errors.
- Made a small unrelated change having to do with how study doc version differences are reported (because referring to headers as "unknown" is confusing and inaccurate).

Files checked in: DataRepo/loaders/msruns_loader.py DataRepo/utils/exceptions.py