Added code to be able to associate a sequence with an mzXML file that is inferred by a peak annotation file existing along the mzXML's path.
Details:

- Created new exception classes:
  - MzxmlNotColocatedWithAnnot
  - MzxmlColocatedWithMultipleAnnot
- Added a property to MSRunSequence: sequence_name to generate a sequence name from a model object.
- Added a default mzxml_dir to load_msruns.py
- Moved around a couple of imports to avoid circular import
- Added a method to the PeakAnnotationFilesLoader: get_dir_to_sequence_dict that generates a dict of directory to list of unique sequences
- Added method get_msrun_sequence_from_dir to MSRunsLoader that takes the mzxml path/file (and the default sequence) and returns a sequence based on the sequence of peak annotation files that exist along the mzXML's path.
- Restructured the get_or_create_msrun_sample_from_mzxml code in order to retrieve the sequence based on the mzxml's directory path.
- Moved some code in the load_data method to the called methods to make the load_data method smaller.  The idea behind what code goes in load_data is that it should call every get_or_create, so that the looping can represent the number of created methods (generally)
- Made the MSRunsLoader's constructor use the PeakAnnotationFilesLoader to retrieve the peak annotation directory to sequence list dict, used to determine the sequence that an mzXML belongs to.

Ideally, we wouldn't have to rely on peak annotation file placement (given the M:M relationship they have with mzXML files.  Ideally, we could put the mzXML directory on the ms data share in the sequences sheet, but there's no automatic way to do that and the researchers may not even have access to that when they're compiling a submission.  Thus, it would be great if the curator could fill this in, because they have access when they're running the loaders.

It might even be cool for the code to automatically locate each peak annotation file, so that the path doesn't have to be added to the spreadsheet.

Thus, it would be great if the load script could generate a modified study doc (to be able to automatically fix these issues).

Files checked in: DataRepo/loaders/msruns_loader.py DataRepo/loaders/peak_annotation_files_loader.py DataRepo/loaders/peak_annotations_loader.py DataRepo/management/commands/load_msruns.py DataRepo/models/msrun_sequence.py DataRepo/utils/exceptions.py