When loading a study doc, look for mzXML and peak annotation files in the directory containing the study doc by default.
While I was working on updating the contrib doc for version 3, I noted that the old commands did not have to change directories into the study directory in order to load it, and I consulted the old loading code and saw that it basically used the directory of the loading.yaml file.  I applied that strategy to the new loader.

Details:

- Fixed little documentation and f-string issues I encountered.
- Renamed a variable in the study loader: loader_class to "derived_class" for accuracy.
- Added a filename return to PeakAnnotationFilesLoader.get_file_and_format when the file is not found.

Files checked in: DataRepo/loaders/peak_annotation_files_loader.py DataRepo/loaders/study_loader.py