Fixed a problem that was prompted by a failed test.
BuildSubmissionView.extracted_exceptions was not getting added to for retrieved errors because it was no longer getting modified on the for loop line.  All such exceptions are always made into warnings, so I just eliminated the check of is_error being false.

Files checked in: DataRepo/views/upload/submission.py