Fixed tests by modifying MSRunsLoader.get_sample_by_name to have a fallback that checks to see if the sample name has dashes.
There are multiple places from which the method is called and how the sample name is massaged is different for different purposes, but mainly when dealing with accounting for differences between the mzXML file names and the sample names.  This branch has a fix that accounts for the edge case where a skipped mzXML was not included in a peak annotation file, and that change revealed an edge case inside get_sample_by_name that was not covered.  This change also accounted for sample name differences involving mzXML files whose names started with a number.  All of these case combinations are now accounted for in the method.

This change is not ideal and a holistic refactor should be applied to make the code simpler, but given the current time constraints, this patch is pragmatic.

Files checked in: DataRepo/loaders/msruns_loader.py DataRepo/utils/exceptions.py