I got the mzXML droparea to be able to print all of the file paths from a dropped folder.
This doesn't appear to be the same as the webkitRelativePath, so I'm not sure this will work with the file input element which uses that relative path...

There's also an issue with promises.  I'm not getting the paths back, but it took awhile to get to this point where SOMETHING works, so I am checking this in.

Files checked in: DataRepo/forms.py DataRepo/templates/submission/includes/1_start.html static/js/file_list_drop_area.js static/js/submission_start.js
[skip ci]