ListView pagination compatible with bootstrap-table search, sort, and export.
This is still just sketched out, but I just need to trigger reloads with all of the parameters, such as offset, limit, search, order, sort, and all of the column filter values.  Then I need to add javascript to re-render those things so that what the user sees makes sense.  E.g. They see the purple arrow showing that the surrent sort is on column X and that column Y is currently filtered, they're on page Z, and when they export, all of the data is triggered to download (not just the current page worth).

Files checked in: DataRepo/templates/DataRepo/archive_file_list.html DataRepo/views/models/archive_file.py
[skip ci]