First pass at turning the changes to the ArchiveFileListView into a derived class.
There's currently a speed issue.  I am not sure why.  The only essential difference that I can think could cause the slow down is the fact that I decided to always annotate the first_study.  My testing in the shell suggests that when you combine that with a search that OR's all the fields together (even though the annotation isn't included in the query), the query just doesn't return when you try to count it.  Checking in so I can compare this to the initial working version before the attempt to use a base class.  According to how I understand how it's supposed to work, I should see the slow-down on the initial version is I do a search when the Study column is sorted.

Details:

- Created base class BootstrapTableListView
- Created class BootstrapTableColumn
- Made ArchiveFileListView inherit from BootstrapTableListView
- Added method to text_utils: camel_to_title.  I have yet to use it for the title of the page.
- Added a few cookie methods in both the view and the javascript to both prepend the view name and also prepend a column name.
- The new classes are well annotated, so I won't go into a lot of detail on them, but basically, the BootstrapTableListView has a columns attribute that is a list of BootstrapTableColumn objects.  Each BootstrapTableColumn object defines pretty much all of the th attributes in the template containing the bootstrap table.

Files checked in: DataRepo/templates/DataRepo/archive_file_list.html DataRepo/utils/text_utils.py DataRepo/views/models/archive_file.py DataRepo/views/models/base.py
[skip ci]