My suspicion mentioned in my last commit message is confirmed.  Does a coalese annotation BEFORE a filter, regardless of whether the annotation is included in the filter is VERY slow.  If I do any search or column filter while the first_study column is sorted, the original code is just as slow as the new code.  I implemented a work-around.
Any coalesce annotation is only ever applied AFTER the filter has been applied and that has restored the speed in that situation.  Luckily, the Coalesce functionality was never intended in my initial design to be included in a filter, because searching the individual related fields does not increase the number of results.  I added a console warning print if a custom annotation (starting) with an outer Coalesce to mention that it will not be included in the search and that the fields will be searched instead.

There are still some debug prints in this commit.

The same logic I used in get_any_field_query to decide whether to use the annotation or field in the search was also applied to the column filter operation.

I also fixed a bug where if you set the rows per page low and go to a high numbered page and then do a search, filter, or rows per page change, you get a page not found error, by changing the page to 1 in any of those events.  You can still create the error by changing the page in the URL/location field to change the page number.  It would be nice if I could handle that more gracefully.

Files checked in: DataRepo/templates/DataRepo/archive_file_list.html DataRepo/views/models/base.py
[skip ci]