Fixed an issue where a dtype dict's keys doesn't match any of the headers and an exception is raised.
This was the case for the protocols sheet (i.e. treatments sheet), so it affected a ton of tests.

I used StudyLoader.CustomLoaderKwargs to grab the default headers.  This doesn't account for custom headers, but that has been an issue for awhile name.  Hopefully, this is enough to fix all the tests.

Files checked in: DataRepo/management/commands/load_study.py