Fixes a bug introduced by commit a8f0c5c which was trying to fix THIS bug.
The problem was that the FCirc format's initial ncmp select list had no default selection.  The fix had been to set the initial value BEFORE updating the choices.  However, the side-effect of that was to always set the first item in the select list.  So whenever a user selected a different value and submited a search, their search parameter for the ncmp select list would be changed.

The actual problem is that the ncmp choices are initially dictated by the first field in the selected template.  That first field in the peak group template is Age (a number), but the first field in the fcirc template is animal (name) (a string).  Therefore, the first item in each ncmp select list ("is") gets the value of "exact" for the number type.  It should be "iexact" for the string type.  The code dynamically updates the choices, but the initial value is stale.

To fix it, I search the ncmp choices after the choices update and if the initial value isn't there, I set the first value.

There has to be a better way to do this, but for now, this works.

Files checked in: static/js/hierarchical_formsets.js