Added the ability for the PeakAnnotationsLoader to remove the compound version numbers that ELMaven appends to compounds sometimes.
Contrary to the branch name, I opted to modify the value in the compound column instead of use the compoundId column, because I have encountered cases where the value in the compoundId column was a generic identifier.

I included the ability to turn off this modification in the method that takes the campounds from the pandas series, but defaulted it to make the modification.

I considered and dismissed the strategy of performing the modification in the pandas dataframe conversion, mainly because we lose control over modifying the behavior easily down the road (e.g. say we encounter a compound ending with "(1)" or some other number legitimately.  I didn't spend much time trying to provide an option to disable it in certain cases, but did this in a way that that's still an option.  The other reason for not doing it in the conversion is because all peak annotation tools take EL Maven data as input, and the purpose of the conversion is to handle variability in the formats.  This modification is the same in every format.

Details:

- Modified PeakAnnotationsLoader.get_peak_group_compounds_dict:
  - Added a fix_elmaven_compounds boolean keyword argument
  - Called fix_elmaven_compound if fix_elmaven_compounds is True
- Added a call to peak_annot_loader.fix_elmaven_compound from DataRepo.views.upload.submission.BuildSubmissionView.extract_autofill_from_peak_annotation_files_forms
- Added tests

Files checked in: DataRepo/loaders/peak_annotations_loader.py DataRepo/tests/loaders/test_peak_annotations_loader.py DataRepo/tests/views/upload/test_submission.py DataRepo/views/upload/submission.py DataRepo/data/tests/data_submission/accucor3_with_elmaven_compound_numbers.xlsx