Added a test for BSTListView.get_context_data.
This time, I decided to use a RequestFactory.  It keeps me from having to set request.method manually, but going this route doesn't setup object_list automatically and I'm not sure I understand why.  I haven't hooked this in yet with the urls, so I can't do a full test.  It is possible that setting the cookies in the constructor is folly.  I think it should work when everything is in place.  I thought it had worked in the prototype.  I'll just have to keep an eye on it and more it if necessary.  The problem is that I don't know what order the methods are executed in.  It might be that I will have to find another method to override.

Files checked in: DataRepo/tests/views/models/bst/test_list_view.py