Non-ideal fix to not reporting missing serum samples when an error has occurred in the samples load (preventing samples from being loaded).
We don't want to needlessly report animals without serum samples that would potentially be resolved by fixing other errors that prevented the loading of those samples.  But unfortunately, in the StudyLoader, it is crafting the errors and summarizing them, which involved removing repeated errors and summarizing them.  They are collected across multuple load scripts and the errors from the samples loader are turned into warnings, so I couldn't use that.  Probably, the packaging of these exceptions should happen after the animal check, in a separate loop.

Files checked in: DataRepo/loaders/study_loader.py