Optimized the query in PeakGroupsToMzxmlZIP.queryset_to_files_iterator.
The hope is here that this leads to mzXML files being downloaded.  There is a bug in the mzXML download feature where only the tsv is in the resulting zip file, but when I ran the exact same code in the shell, it seemed to be running as expected, just veeeeeeery slowly.  I'm taking a guess that what is happening is some sort of timeout, because after it got to all files, the loop just furiously spun along even though it had already accumulated all of the files.  This spinning occurs because it repeatedly encounters files it has already added to the zip and that the stop in responding of the server prematurely ends the download (without error)?

If this doesn't fix the download, hopefully, it will speed things up.  In the shell, this change went from unended looping to a completion of the loop in a handful of seconds (2, maybe 3).

Files checked in: DataRepo/views/search/download.py