Switched back to top-aligned headers.
The reason I had experimented with bottom-alignment was that all the filter fields were unevenly aligned, but I figured out that that was because of a valign-top option in the th attributes.  Once I removed that, it started rendering correctly.

One caveat though is that I still had to include a CSS class for th-inner as an override, because adding position relative to the existing css didn't work.  Not sure why.

Files checked in: DataRepo/static/css/bst.css DataRepo/tests/templates/models/bst/test_th.py