`TraceBaseDocs` `Upload` Updates (#1656)
## What

- Resolves
[GREATS-10](https://princeton-university.atlassian.net/browse/GREATS-10)
- Resolves #1645
- Resolves #1441
- Resolves #800
- Resolves #755

## How

See the **ISSUE OWNER SECTION** of issue #1645.

Details:

- Updated the upload section of the user-facing documentation.
- Renamed the "Glossary of Calculated Values" file/title to put it at
the top of the values section.
- Added a note about markdownlint in `CONTRIBUTING.md` since it was
giving me a deprecation warning, which went away after I updated it to
the latest version and it's not a part of the requirements docs.
- Fixed a new `markdownlint` error from the newer linter version in `How
to label sample information.md` about the table needing to be surrounded
by empty lines.
  - Added some updated/new screenshots used in `How to Upload.md`
  - Made updates to `How to Upload.md` to reflect:
    - How v3 can hide/disable the upload tab
    - Show a new screenshot that replaces the "browse" tab with "home".
  - Changed dashed bullets with asterisks for `mkdocs` requirements.
- Started to change the bulleted step-by-step instructions with numbered
subheaders so that the steps can be more descriptive.
  - Added a preamble to encourage more granular submissions.
- Described the overall submission process before getting into the
step-by-step instructions.
- Fleshed the "start" tab with new submission process details and a
screenshot example.
- Updated "How to Upload" and created "Uploading FAQ". Added a
screenshot and fixed an image URL.
- Organized, streamlined, fleshed, and updated the upload documentation
for v3 of tracebase.
- Fixed text/markdown linting issues and added documentation and
settings changes for the `textlint` linter.
- Added documentation about errors and warnings that can occur during
submission building and that was generated by sphinx run on the
exceptions file in the main branch (see PR #1655) and manually edited it
to be user-facing.
- Polished the new content and reviewed how everything displays. 
- I noted that a linter comment at the top of the fill-in page was
affecting how the page was presented in the left panel, so I moved the
comment under the top header. It was causing `2` to show up in the link.
- Removed duplicate exceptions from the reference. Note sure why those
appeared.
  - Made all exception headings be wrapped in code quotes.
  - Removed some irrelevant/empty exceptions I found.
- Fixed up the display of some non-markdown docstring content that was
not rendering well. This is due to the docstrings containing a mix of
markdown and other common docstring formatting.
- Made sure all sheet and column references were wrapped in code quotes.
- Fixed other linter comments that were affecting how the content was
line-broken.

## Why

See the **Feature Inspiration** section of issue #1645.

## Tests

Testing was performed via superlinter CI tests and via the procedure
described in the **Use mkdocs to generate a local static site** section
of `TraceBaseDocs/README.md`. All errors and warnings generated by
`mkdocs` were fixed and every page was manually browsed to look for
malformed content (such as improper unordered list indenting).

## Security Concerns

None

---

## Reviewer Notes/Requests

None

## Checklist

- Review Requirements
  - [x] Minimum approvals: 1
  - [x] No changes requested
  - [x] All blocking issues resolved by reviewers
  - [x] Specific reviewers: None
  - Review period: 2 days
- Associated Issue/PR Requirements:
  - [x] All issue requirements are satisfied
- The issue requirements for the about page and splash page updates
(`1645.6.` and `1645.7.`) will be a part of a future effort.
  - [x] All PR dependencies are merged
- Basic Requirements
- [x] [Linters
pass](https://github.com/Princeton-LSI-ResearchComputing/tracebase/blob/main/CONTRIBUTING.md#linting)
- [x] [Tests
pass](https://github.com/Princeton-LSI-ResearchComputing/tracebase/blob/main/CONTRIBUTING.md#quality-control)
  - [x] Conflicts resolved
- Overhead Requirements
- [x] [Tests
implemented/updated](https://github.com/Princeton-LSI-ResearchComputing/tracebase/blob/main/CONTRIBUTING.md#test-implementation)