3.1.6 Manuscript Release (#1657)
Copied the git ignore file from main to avoid accidentally checking in my scratch files I started using.

Renamed some links in the TraceBaseDocs/index.md file for brevity and made some draft updates to the download docs.

I moved some linked doc contents into index, since they were very small files.

Also removed a TODO comment for an analysis example that was never created and added a new example I found on Lance's rpubs.com page for FCirc bar plots.

Download and analysis documentation updates. 

Details:

- Updated screenshots.
- Edited the index.md file for brevity.
- Consolidated redundant download format info.
- Changed external rpubs.com example links to static html files (generated in RStudio).
- Added information about the mzXML Zip download format.
- Updated relative links throughout the documentation for the directory reorganization.

Improved the precision of the FCirc definition as described in issue #634, with liberties.

Details:

- Added a screenshot showing how to get the most accurate values.
- Removed external image link by copying the image into the attachments directory.

Updated the documentation in the Overview directory.
Eliminated the terms and definitions doc, as required in issue 801.

Updated the Overview directory documents and renamed Overview to "About".

Details:

- Updated Overview links to use "About".
- Edited the slideshow link in the index (under the overview section) to be more succinct.
- Fixed non-working anchor links throughout.  The way anchor links were originally set up relied on an ununstalled mkdocs dependency.
- Made general updates for version 3.
- Updated the broken developer LSI staff links.
- Added details on how to perform an advanced search.
- Linked the varying hierarchical institution links (PU, LSI, Research Computing, and Rabinowitz lab) on the Team page.
- Added 2 sections: TraceBase Design and Under the Hood as a sort of primer on the guiding principles overseeing TraceBase development and all of the hidden features that ensure smooth and accurate running of TraceBase.

Changed LSI people links to email addresses.

Updated the TraceBaseDocs documentation for version 3, specifically for the files under the Values directory.

Updated the documentation in the Overview directory.

Eliminated the terms and definitions doc, as required in issue 801.

Updated the Overview directory documents and renamed Overview to "About".

Details:

- Updated Overview links to use "About".
- Edited the slideshow link in the index (under the overview section) to be more succinct.
- Fixed non-working anchor links throughout.  The way anchor links were originally set up relied on an ununstalled mkdocs dependency.
- Made general updates for version 3.
- Updated the broken developer LSI staff links.
- Added details on how to perform an advanced search.
- Linked the varying hierarchical institution links (PU, LSI, Research Computing, and Rabinowitz lab) on the Team page.
- Added 2 sections: TraceBase Design and Under the Hood as a sort of primer on the guiding principles overseeing TraceBase development and all of the hidden features that ensure smooth and accurate running of TraceBase.
- Changed LSI people links to email addresses.

Details:

- Updated the upload section of the user-facing documentation.
- Renamed the "Glossary of Calculated Values" file/title to put it at
the top of the values section.
- Added a note about markdownlint in `CONTRIBUTING.md` since it was
giving me a deprecation warning, which went away after I updated it to
the latest version and it's not a part of the requirements docs.
- Fixed a new `markdownlint` error from the newer linter version in `How
to label sample information.md` about the table needing to be surrounded
by empty lines.
  - Added some updated/new screenshots used in `How to Upload.md`
  - Made updates to `How to Upload.md` to reflect:
    - How v3 can hide/disable the upload tab
    - Show a new screenshot that replaces the "browse" tab with "home".
  - Changed dashed bullets with asterisks for `mkdocs` requirements.
- Started to change the bulleted step-by-step instructions with numbered
subheaders so that the steps can be more descriptive.
  - Added a preamble to encourage more granular submissions.
- Described the overall submission process before getting into the
step-by-step instructions.
- Fleshed the "start" tab with new submission process details and a
screenshot example.
- Updated "How to Upload" and created "Uploading FAQ". Added a
screenshot and fixed an image URL.
- Organized, streamlined, fleshed, and updated the upload documentation
for v3 of tracebase.
- Fixed text/markdown linting issues and added documentation and
settings changes for the `textlint` linter.
- Added documentation about errors and warnings that can occur during
submission building and that was generated by sphinx run on the
exceptions file in the main branch (see PR #1655) and manually edited it
to be user-facing.
- Polished the new content and reviewed how everything displays. 
- I noted that a linter comment at the top of the fill-in page was
affecting how the page was presented in the left panel, so I moved the
comment under the top header. It was causing `2` to show up in the link.
- Removed duplicate exceptions from the reference. Note sure why those
appeared.
  - Made all exception headings be wrapped in code quotes.
  - Removed some irrelevant/empty exceptions I found.
- Fixed up the display of some non-markdown docstring content that was
not rendering well. This is due to the docstrings containing a mix of
markdown and other common docstring formatting.
- Made sure all sheet and column references were wrapped in code quotes.
- Fixed other linter comments that were affecting how the content was
line-broken.